All visible bugs fixed

+++ src/main/java/com/fivesysdev/Fiveogram/services/CommentServiceImpl.java
@@ -4,28 +4,23 @@ import com.fivesysdev.Fiveogram.models.Comment;
 import com.fivesysdev.Fiveogram.models.Post;
 import com.fivesysdev.Fiveogram.repositories.CommentRepository;
 import com.fivesysdev.Fiveogram.serviceInterfaces.CommentService;
-import com.fivesysdev.Fiveogram.serviceInterfaces.PostService;
+import com.fivesysdev.Fiveogram.serviceInterfaces.UserService;
 import com.fivesysdev.Fiveogram.util.Context;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.Map;
 
 @Service
 @Transactional
 public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
-    private final PostService postService;
+    private final UserService userService;
 
-    public CommentServiceImpl(CommentRepository commentRepository, PostService postService) {
+    public CommentServiceImpl(CommentRepository commentRepository, UserService userService) {
         this.commentRepository = commentRepository;
-        this.postService = postService;
-    }
-
-    @Override
-    public void addComment(Post post, Comment comment) {
-        post.addComment(comment);
-        commentRepository.save(comment);
+        this.userService = userService;
     }
 
     @Override
@@ -34,26 +29,38 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void editComment(long id, String text) {
-        Comment comment = createComment(id, text);
-        commentRepository.save(comment);
+    public Map<String, String> editComment(long id, String text) {
+        Comment oldComment = commentRepository.findCommentById(id);
+        if (oldComment == null) {
+            return Map.of("Message", "Comment not found");
+        }
+        if (oldComment.getAuthor().equals(userService.getUser(Context.getUserFromContext().getId()))) {
+            oldComment.setText(text);
+            return Map.of("Message", "ok");
+        }
+        return Map.of("Message", "That`s not your comment");
     }
 
     @Override
-    public void deleteComment(long id) {
-        commentRepository.deleteById(id);
+    public Map<String, String> deleteComment(long id) {
+        Comment oldComment = commentRepository.findCommentById(id);
+        if (oldComment == null) {
+            return Map.of("Message", "Comment not found");
+        }
+        if (oldComment.getAuthor().equals(userService.getUser(Context.getUserFromContext().getId()))
+                || oldComment.getPost().getAuthor().equals(userService.getUser(Context.getUserFromContext().getId()))) {
+            commentRepository.deleteById(id);
+            return Map.of("Message", "ok");
+        }
+        return Map.of("Message", "That`s not your comment");
     }
 
-    public boolean removeComment(Comment comment) {
-        return true;
-    }
-
-    public Comment createComment(long id, String text) {
+    public Comment createComment(Post post, String text) {
         Comment comment = new Comment();
         comment.setAuthor(Context.getUserFromContext());
         comment.setText(text);
-        comment.setPublished((LocalDate.now()));
-        comment.setPost(postService.findPostById(id));
+        comment.setPublished((LocalDateTime.now()));
+        comment.setPost(post);
         return comment;
     }
 }
